Remove the lock to Hubs Foundation master
What: Removes the if guard from the publish step in the publish workflow file that prevents publishing on all but the Hubs Foundation's master branch.

Why: This isn't needed because this workflow only publishes to GitHub, so there are no issues with forks running it, and it is useful to publish branches other than master because it gives you an easily distributable/installable zip (only part of this repository is the actual addon, so that part needs to be separated out into its own zip to be installed into Blender, which the workflow does along with adding the workflow run number onto the version number to provide a semi-unique identifier).